Don't include termio.h if compile with android toolchain, since current bionic C does not have this header. __ANDROID__ is defined by androideabi toolchain.
Currently, build of openssl includes termios.h, instead of
termio.h. This patch makes sure that we keep this behavior
before/after we upgrade Android toolchain.

In this patch, we also update the patch file and config file,
such that this patch is not dropped when openssl is upgraded
in the future.

Change-Id: I35e8dcce456966f585ff07b5ff4edf7ee4cf8baf