Remove SHA0 from openssl
Recent a bug was found that would have been much more obvious if not
for the confusion that "sha" means "SHA-0" in openssl and "SHA" means
"SHA-1" to Java programmers. Removing SHA-0 should be not be an
interoperability issue, it was never really used, was apparently
flawed, so like MD2 we will just remove it.

Bug: 2997009

Change-Id: I630c851fb2f5f344ef7a2c62c7092843cb40818c